backend/requirements.txt
flask
flask-marshmallow
Flask-Bcrypt
flask-cors
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Caching
celery
redis
eventlet
requests
reportlab